docker/requirements.txt
docker==6.1.3
PyYAML==6.0.1
requests==2.31.0
redis==4.6.0